move tests for sorted multi map
